fix: do not delete mols when cache is enabled
- add prop to chose caching options

See #20 #19

+++ src/utils/molecule.ts
@@ -32,7 +32,9 @@ export const get_molecule_details = ({ smiles }: GetMoleculeDetailsParams, RDKit
   const mol = get_molecule(smiles, RDKit);
   if (!mol) return null;
   const details = JSON.parse(mol.get_descriptors());
-  mol.delete();
+  if (!globalThis.rdkitProviderGlobals.jsMolCacheEnabled) {
+    mol.delete();
+  }
 
   return {
     numAtoms: details.NumHeavyAtoms,
